Phase 2: reorganize into core/, ines/, inev/, prepp/, kraken/ modules
Move files into logical subpackages and update all imports:
- core/: shared infrastructure (node, network, graph, structures, etc.)
- ines/: INES processing pipeline (combigen, projections, selectivity, etc.)
- inev/: INEv graph construction (placement_aug, process_combination, filter)
- prepp/: PrePP push-pull optimization (prepp, generate_eval_plan)
- kraken/: renamed from kraken2_0/ (internal structure unchanged)
