+++ src/core/all_pairs.py
