+++ src/core/parse_network.py
