+++ src/core/proj_string.py
