+++ src/kraken/search/greedy.py
